requirements.txt
pandas==2.2.3
requests==2.32.3
BeautifulSoup4==4.12.3
python-dotenv==1.0.1
pyodbc==5.1.0
openpyxl==3.1.5
xlrd==2.0.1